+++ drivers/dma/CMakeLists.txt
@@ -7,7 +7,7 @@ zephyr_library_sources_ifdef(CONFIG_DMA_STM32		dma_stm32.c)
 zephyr_library_sources_ifdef(CONFIG_DMA_STM32_V1	dma_stm32_v1.c)
 zephyr_library_sources_ifdef(CONFIG_DMA_STM32_V2	dma_stm32_v2.c)
 zephyr_library_sources_ifdef(CONFIG_DMAMUX_STM32	dmamux_stm32.c)
-zephyr_library_sources_ifdef(CONFIG_DMA_DW		dma_dw.c)
+zephyr_library_sources_ifdef(CONFIG_DMA_DW		dma_dw.c dma_dw_common.c)
 zephyr_library_sources_ifdef(CONFIG_DMA_NIOS2_MSGDMA	dma_nios2_msgdma.c)
 zephyr_library_sources_ifdef(CONFIG_DMA_SAM0		dma_sam0.c)
 zephyr_library_sources_ifdef(CONFIG_USERSPACE		dma_handlers.c)
