+++ drivers/dma/Kconfig.dw
@@ -1,4 +1,4 @@
-# CAVS DMA configuration options
+# DesignWare DMA configuration options
 
 # Copyright (c) 2018 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
